[Python API] Add missing methods to Datatype/Term
Fixes #4942. The Python API was missing some methods related to
datatypes. Most importantly, it was missing `mkDatatypeSorts`, which
meant that datatypes with unresolved placeholders could not be resolved.
This commit adds missing methods and ports the corresponding tests of
`datatype_api_black.h` to Python. The commit also adds support for
`__getitem__` in `Term`.